2021.03.01 by zxz

+++ src/main/java/net/zlw/cloud/designProject/mapper/ProjectMapper.java
@@ -980,7 +980,7 @@ public interface ProjectMapper extends Mapper<BaseProject> {
                     "s2.status = '6'"
 
     )
-    Double VisaChangeCount(CostVo2 costVo2);
+    Integer VisaChangeCount(CostVo2 costVo2);
 
     @Select(
             "SELECT " +
